chore: move client tests to client_test.go
